L-Variant/consumers/influx/requirements.txt
influxdb~=5.3.1
kafka-python~=1.4.7